AppointmentsApp/Backend/requirements.txt
Flask
Flask-JWT-Extended
firebase-admin
flask-limiter
flask-cors
flask-talisman
cryptography
google-cloud-storage
setuptools
pyrebase4
pytest